Conflicts:
	utils/scannercontrol.cpp
	utils/scannercontrol.h

+++ utils/scannercontrol.h
@@ -17,7 +17,7 @@ public slots:
     bool openDevice(QString name);
     bool laserOn();
     bool laserOff();
-    bool step(qreal dist);
+    bool step(qint64 dist);
     int getPos();
     int getRealPos();
     void close();
